// src/enums/cache-tags-enum.ts
export const CacheTagsEnum = {
  Transaction: "transaction",
  FinancialAccount: "financialAccount",
  Card: "card",
  Payment: "payment",
  Customer: "customer",
  CustomerCharges: "customerCharges",
  CustomerInvoices: "customerInvoices",
  Invoice: "invoice",
  Schedule: "schedule",
  ScheduleProject: "scheduleProject",
  Account: "account",
  Service: "service",
  Event: "event",
  Tax: "tax",
  Whoami: "whoami",
  Business: "business",
  FinancialAccountDashboard: "financialAccountDashboard",
  Subscription: "subscription",
  Charge: "charge",
};

/* 
export enum CacheTagsEnum {
    Transaction = "transaction",
    FinancialAccount = "financialAccount",
    Card = "card",
    Payment = "payment",
    Customer = "customer",
    Invoice = "invoice",
    Schedule = "schedule",
    ScheduleProject = "scheduleProject",
    Account = "account",
    Service = "service",
    Event = "event",
    Tax = "tax",
    Whoami = "whoami",
    Business = "business",
    FinancialAccountDashboard = "financialAccountDashboard",
    Subscription = "subscription",
  } */